chore: code minor reorganization
