+++ bin/delete-photos/PhotoDeleter.ts
@@ -1,13 +1,19 @@
+import { DeleteResult } from 'mongodb';
 import { PhotoId } from '../../src/models/Photo';
 import { CommandStatus } from './CommandStatus';
-import {
-  DeleteFileFromStorage,
-  DeleteFilesResponse,
-  DeletePhotosById,
-  GetPhotosToDelete,
-} from './services';
+
+export type DeletePhotosById = (ids: Array<PhotoId>) => Promise<DeleteResult>;
+export type DeleteFileFromStorage = (id: Array<PhotoId>) => Promise<DeleteFilesResponse>;
+export type GetPhotosToDelete = (limit: number) => Promise<Array<PhotoId>>;
 
 type PhotosIdChucks = Array<Array<PhotoId>>;
+
+export type DeleteFilesResponse = {
+  message: {
+    confirmed: string[];
+    notConfirmed: string[];
+  };
+};
 
 export class PhotoDeleter {
   private readonly status: CommandStatus;
@@ -27,9 +33,6 @@ export class PhotoDeleter {
 
     this.status.updateRequest(results);
 
-    console.log(results);
-    console.log('SOME REJECTEC', results.filter(result => result.status === 'rejected').length);
-
     const photosIdsRemoved = results
       .filter((r) => r.status === 'fulfilled')
       .flatMap((r) => (r as PromiseFulfilledResult<DeleteFilesResponse>).value.message.confirmed);
@@ -39,17 +42,17 @@ export class PhotoDeleter {
     this.status.updatePhotosRemoved(photosIdsRemoved);
 
     return photosIdsRemoved;
-  };
+  }
 
   public async run(limit: number, concurrency: number) {
     this.status.init();
     try {
-      const chunksOf = Math.ceil(limit / concurrency);
+      const numberOfChunks = Math.ceil(limit / concurrency);
 
       let ids = await this.getPhotosIdsToDelete(limit);
 
       while (ids.length > 0) {
-        const chuncks = Array.from(new Array(chunksOf), (_, i) =>
+        const chuncks = Array.from(new Array(numberOfChunks), (_, i) =>
           ids.slice(i * concurrency, i * concurrency + concurrency),
         );
